Added the MJ12Bot to the bad list
This crawler is supposed to be respectful. Unfortunately it ties Magento's core_url_rewrite table in knots by hammering non-existent links with non-existant parameters. More to the point, this crawler exists only so that Majestic SEO can sell the data it generates. It'll never be of any use to anyone that doesn't hire them, so blocking it is fine.